init api rsp fix

+++ app/src/main/java/com/mesum/library/data/ListPriceX.kt
@@ -1,6 +1,6 @@
 package com.mesum.library.data
 
 data class ListPriceX(
-    val amountInMicros: Int,
+    val amountInMicros: Double,
     val currencyCode: String
 )
