+++ app/src/main/java/com/mesum/library/data/RetailPrice.kt
@@ -1,6 +1,6 @@
 package com.mesum.library.data
 
 data class RetailPrice(
-    val amountInMicros: Int,
+    val amountInMicros: Double,
     val currencyCode: String
 )
